@@ -1,22 +1,22 @@
-const { contextBridge, ipcRenderer } = require('electron');
+import { contextBridge, ipcRenderer } from "electron";
 
 // Expose protected methods that allow the renderer process to use
 // the ipcRenderer without exposing the entire object
 contextBridge.exposeInMainWorld(
 	'ipcRenderer', {
-		send: (channel, ...data) => {
+		send: (channel: string, ...data: any[]) => {
 			const validChannels = ['hide', 'height'];
 			if (validChannels.includes(channel)) {
 				ipcRenderer.send(channel, ...data);
 			}
 		},
-		sendSync: (channel, ...data) => {
+		sendSync: (channel: string, ...data: any[]) => {
 			const validChannels = ['find', 'perform', 'resolve'];
 			if (validChannels.includes(channel)) {
 				return ipcRenderer.sendSync(channel, ...data);
 			}
 		},
-		receive: (channel, func) => {
+		receive: (channel: string, func: (...args: any[]) => void) => {
 			const validChannels = ['show', 'blur'];
 			if (validChannels.includes(channel)) {
 				// Deliberately strip event as it includes `sender` 
